start exchange

+++ src/com/mfm_wallet/model/Candle.java
@@ -4,9 +4,9 @@ public class Candle {
     public String key;
     public String period_name;
     public long period_time;
-    public float low;
-    public float high;
-    public float open;
-    public float close;
-    public float value;
+    public double low;
+    public double high;
+    public double open;
+    public double close;
+    public double value;
 }
